feat(renderer): basic antialiased segment (fix color calculations)

+++ Color.h
@@ -10,7 +10,7 @@ class Color {
 public:
     static color_int fromRGBW(uint8_t r, uint8_t g, uint8_t b, uint8_t w);
     static color_int fromRGB(uint8_t r, uint8_t g, uint8_t b);
-    static color_int fromHSV(uint16_t h, uint16_t s, uint16_t v);
+    static color_int fromHSV(double h, double s, double v);
 };
